==> notebooks/1 DataPrep-SQL/2 VariableMapping.ipynb <==
import pandas as pd
import numpy as np

# Inputs

hist_lls_ftable_csvfile = None
hist_lr_ftable_csvfile = 'D:/Data/linkedclaims_casuals_2018m04_hist_lr_ftable.csv'

# Outputs

mapping_csvfile = 'D:/Data/linkedclaims_casuals_2018m04_variables_mapping.csv'

# Map Variable Values

def remove_contiguous_duplicates_from_list(values, null_values = ['', np.NaN]):
    ret = []
    values = [x for x in values if x not in null_values]
    for i in range(0, len(values)-1):
        if values[i] != values[i+1] and values[i]:
            ret.append(values[i])
    try:
        ret.append(values[len(values)-1])
        return ret
    except:
        return ['']
    
def remove_contiguous_duplicates_from_list_2(values, null_values = ['', np.NaN]):
    if len (values) == 3:
        return values
    
    ret = []
    spare = []
    values = [x for x in values if x not in null_values]
    if np.mod(len(values),2) != 0:
        spare.append(values[len(values)-1])
        values = values[0:len(values)-1]
    
    for i in range(0, int(len(values)), 2):
        myvals = [values[i], values[i+1]]
        if len(ret) == 0:
            ret.extend(myvals)
        else:
            found = False
            for i in range(0, int(len(ret)), 2):
                if myvals == [ret[i], ret[i+1]]:
                    found = True
                    break
            if found == False:
                ret.extend(myvals)
    
    if len (spare) > 0:
        ret.extend(spare)
    return ret
    
def concat_values_from_list(values, concatenator = '-', null_values = ['',np.NaN]):
    concat = ''
    for elem in values:
        if elem not in null_values:
            if len(concat) == 0:
                concat = str(elem)
            else:
                concat = concat + concatenator + str(elem)
    return concat
    
def value_to_simple (value, splitter = '-'):
    svalue = value.replace('C-', 'C')
    svalues = svalue.split(splitter)
    svalues = remove_contiguous_duplicates_from_list(svalues)
    svalues = remove_contiguous_duplicates_from_list_2(svalues)
    svalue = concat_values_from_list(svalues)
    return svalue

def create_mapping(variable, original_values, target_value):
    mapping = pd.DataFrame()
    mapping['OriginalValue'] = original_values
    mapping['Variable'] = [variable]*len(original_values)
    mapping['TargetValue'] = [target_value]*len(original_values)
    return mapping

mapping = pd.DataFrame()

################# hist_lls
if hist_lls_ftable_csvfile is not None:
    hist_lls_ftable = pd.read_csv(hist_lls_ftable_csvfile)
    variable = 'hist_lls'
    original_values = hist_lls_ftable[variable].tolist()
    for elem in original_values:
        target_value = value_to_simple(elem)
        if elem != target_value:
            from_values = []
            from_values.append(elem)
            df = create_mapping(variable, from_values, target_value)
            mapping = mapping.append(df, ignore_index=True)
        

################# hist_lr
if hist_lr_ftable_csvfile is not None:
    hist_lr_ftable = pd.read_csv(hist_lr_ftable_csvfile)
    variable = 'hist_lr'
    original_values = hist_lr_ftable[variable].tolist()
    for elem in original_values:
        target_value = value_to_simple(elem)
        if elem != target_value:
            from_values = []
            from_values.append(elem)
            df = create_mapping(variable, from_values, target_value)
            mapping = mapping.append(df, ignore_index=True)


################# NAT_CODE
variable = 'nat_code'
original_values = ["000","007"]
target_value = "Irish"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'nat_code'
original_values = ["006", "013", "018", "019", "020"]
target_value = "UK"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'nat_code'
original_values = ["001","002","003","004","005","008",
                   "009","010","011","012","030","032",
                   "038"]
target_value = "EU13"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'nat_code'
original_values = ["060","064","066","068","046",
                   "191","233","703","705","428","440",
                   "203","600"]
target_value = "EU15_EU28"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'nat_code'
original_values = ['014','015','016','024','028','031',
                   '036','043','048','051','052','070','071',
                   '072','112','158','195','198','204','208',
                   '214','216','220','222','228','232','236',
                   '240','244','247','248','252','257','260',
                   '264','268','272','276','280','284','288',
                   '302','306','310','311','314','318','322',
                   '324','328','330','334','342','346','350',
                   '352','355','366','370','373','375','378',
                   '382','386','389','391','393','395','398',
                   '400','404','410','412','413','416','417',
                   '421','424','430','432','436','442','448',
                   '452','453','456','458','460','463','464',
                   '465','467','469','472','473','480','484',
                   '488','492','498','500','504','508','512',
                   '516','520','524','528','554','604','608',
                   '612','616','624','628','632','636','640',
                   '643','647','649','660','662','664','666',
                   '667','669','672','675','676','680','690',
                   '696','700','701','706','708','710','716',
                   '720','732','736','740','743','750','762',
                   '800','801','804','806','812','815','817',
                   '860','887','890','892','897','902','903',
                   '905','907','913','915','920','921','924']
target_value = "Other"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'nat_code'
original_values = ["."]
target_value = ""
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

################# sex
variable = 'sex'
original_values = ["W","F","w","f"]
target_value = 2
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'sex'
original_values = ["M","m"]
target_value = 1
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)


################# marital_status
variable = 'marital_status'
original_values = ["M","E"]
target_value = "Married"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'marital_status'
original_values = ["B","I","C","V","S","N","U","Z","P","D","L","T"]
target_value = "Single"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'marital_status'
original_values = ["W","H"]
target_value = "Widowed"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

################# ada_code
variable = 'ada_code'
original_values = ["A","B"]
target_value = "SpouseOnSocialWelfare"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["C", "H", "J"]
target_value = "SpouseNoIncome"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["D", "S"]
target_value = "NoSpouse_SingleSeparatedDivorcedWidowed"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["E"]
target_value = "SpouseEarnings_bigger_than_Euro_310"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["F"]
target_value = "SpouseEarnings_between_Euro_310_400"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["G"]
target_value = "SpouseEarnings_bigger_than_Euro_400"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["P"]
target_value = "SpouseEarnings_less_than_Euro_100"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["T"]
target_value = "SpouseEarnings_between_Euro_100_310"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["U","V","X","Y"]
target_value = "SpouseUnknown"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["W"]
target_value = "SpouseHasEarningsPreviously_E_P_T"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'ada_code'
original_values = ["",".","NaN","nan"]
target_value = "NoDetails" ;
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)


################# location
variable = 'location'
original_values = ["000518LO","000533LO","000535LO","000536LO","000537LO",
                   "000538LO","000539LO","000540LO","000660LO"]
target_value = "CorkCentral"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000554LO","000547LO","000624LO","000625LO","000626LO",
                   "000630LO","000638LO","000639LO","000652LO"]
target_value = "DublinCentral"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000549LO","000628LO","000631LO","000633LO",
                   "000657LO","000658LO"]
target_value = "DublinNorth"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000551LO","000569LO","000627LO","000629LO","000632LO"]
target_value = "DublinSouth"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000647LO"]
target_value = "DublinSouthFacilitatorsOffice"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000697LO"]
target_value = "GuardiansPensions"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000970LO"]
target_value = "MellowsRoadCWO"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000501LO","000502LO","000503LO","000507LO","000550LO","000552LO",
                   "000553LO","000579LO","000598LO","000599LO","000601LO","000602LO","000603LO"]
target_value = "MidlandsNorth"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000504LO","000505LO","000651LO","000506LO","000508LO","000527LO","000528LO","000529LO","000531LO",
                   "000532LO","000584LO","000585LO","000587LO","000596LO","000597LO","000637LO"]
target_value = "MidlandsSouth"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000522LO","000523LO","000524LO","000556LO","000557LO",
                   "000559LO","000570LO","000571LO","000572LO"]
target_value = "MidLeinster"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000573LO","000575LO","000576LO","000577LO","000604LO","000580LO","000594LO","000595LO"]
target_value = "MidWest"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000654LO","000656LO","000710LO","000936LO","000937LO"]
target_value = "National"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000561LO","000562LO","000563LO","000564LO","000565LO","000566LO","000567LO","000544LO","000545LO"]
target_value = "NorthEast"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000568LO","000588LO","000589LO","000590LO","000591LO","000592LO",
                   "000593LO","000525LO","000605LO","000606LO","000607LO","000608LO","000634LO"]
target_value = "NorthWest"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000996LO"]
target_value = "PSCCentreGandonHouse"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000993LO"]
target_value = "PSCCentreGalway"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000808LO"]
target_value = "LODeactivatedOfficers"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000659LO"]
target_value = "RoscommonControlOffice"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000530LO","000615LO","000616LO","000619LO","000620LO","000621LO","000622LO","000666LO"]
target_value = "SouthEast"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000515LO","000516LO","000519LO","000520LO","000521LO",
                   "000534LO","000541LO","000542LO","000543LO","000609LO",
                   "000610LO","000611LO","000612LO","000613LO","000614LO","000635LO"]
target_value = "SouthWest"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000500LO","000581LO","000583LO","000509LO","000510LO",
                   "000511LO","000512LO","000513LO","000514LO","000526LO","000578LO","000618LO"]
target_value = "West"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000661LO"]
target_value = "OnlineApplication"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000662LO"]
target_value = "FrontOrBackOffice"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["000681LO"]
target_value = "BackToWork"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'location'
original_values = ["SWA"]
target_value = "NULL"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

################# Occupation
variable = 'occupation'
original_values = ['010', '011', '012', '013', '014', '015',
                   '016', '017', '018', '019']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "ManagersAndAdministrators"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['020', '021', '022', '023', '024', '025',
                   '026', '027', '028', '029']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "ProfessionalOccupations"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['030', '031', '032', '033', '034', '035',
                   '036', '037', '038', '039']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "AssociateProfessionalAndTechnicalOccupations"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['040', '041', '042', '043', '044', '045',
                   '046', '047', '048', '049']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "ClericalAndSecretarialOccupations"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['050', '051', '052', '053', '054', '055',
                   '056', '057', '058', '059']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "CraftAndRelatedOccupations"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['060', '061', '062', '063', '064', '065',
                   '066', '067', '068', '069']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "PersonalAndProtectiveServiceOccupations"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['070', '071', '072', '073', '074', '075',
                   '076', '077', '078', '079']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "SalesAndCustomerServiceOccupations"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['080', '081', '082', '083', '084', '085',
                   '086', '087', '088', '089']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "PlantAndMachineOperatives"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['090', '091', '092', '093', '094', '095',
                   '096', '097', '098', '099']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "OtherOccupations"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

variable = 'occupation'
original_values = ['100']
original_values = [str(int(x))+'.0' for x in original_values]
target_value = "UnknownOrNotStatedOccupationOrThoseWhoNeverWorked"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)

################# County
variable = 'county'
original_values = ["SWA"]
target_value = "NULL"
df = create_mapping(variable, original_values, target_value)
mapping = mapping.append(df, ignore_index=True)



################# SAVE TO CSV

mapping.to_csv(mapping_csvfile, index=False)

print ('\nALL DONE')
