==> notebooks/1 DataPrep-SQL/3 JLD SAS 2 CSV.ipynb <==
from sas7bdat import SAS7BDAT
import csv
import pandas as pd
import datetime as dt
import time

# Inputs/Outputs

# To convert SAS into CSV
sasfilename = 'D:/Data/linkedclaims_casuals_2018m04.sas7bdat'
csvfilename = sasfilename.replace('.sas7bdat','.csv')

# Convert SAS into CSV

# initiate timing
start_time = time.time()

batch_size = 200000
with SAS7BDAT(sasfilename) as sasfile:
    with open(csvfilename, 'w', newline='') as csvfile:
        wr = csv.writer(csvfile)
        processed = 0
        count = 0
        row_batch = []
        
        eof = 0
        sasfile_iter = sasfile._iter
        while eof == 0:
            row = next(sasfile_iter, None)
            if row is not None:
                row_batch.append(row)
                count = count + 1
            else:
                eof = 1
            if eof or count == batch_size:
                wr.writerows(row_batch)
                processed = processed + count
                print ('Rows processed = %d' %processed)
                count = 0
                row_batch = []
                
elapsed_time = time.time() - start_time
print ('Elapsed time: '+ time.strftime("%H:%M:%S", time.gmtime(elapsed_time)))

print ('\nALL DONE')
